Update the delete user functionality

+++ src/app/users/users.component.ts
@@ -7,31 +7,33 @@ import { DataService } from '../data.service';
   styleUrls: ['users.component.scss']
 })
 
-export class UsersComponent implements OnInit{
+export class UsersComponent implements OnInit {
   users;
   user;
 
-  constructor(private dataService:DataService){}
+  constructor(private dataService: DataService) { }
 
-  ngOnInit(){
+  ngOnInit() {
     this.showusers()
   }
 
 
-  showusers(){
+  showusers() {
     this.dataService.getUsers().subscribe(data => {
       this.users = data;
     })
   }
 
-  deleteUser(id){
-    this.dataService.deleteUser(id).subscribe(data => {
-      console.log(data)
-      this.showusers()
-    })
+  deleteUser(id, name) {
+    if (confirm(`Are you sure you want to delete ${name}?`)) {
+      this.dataService.deleteUser(id).subscribe(data => {
+        this.showusers()
+      })
+    }
+
   }
 
-  updateUser(id, user){
+  updateUser(id, user) {
     //let user = {name: "Edited Name",email: 'ram1@gmail.com', contact: '0000000000'}
     user.name = "Edited"
     this.dataService.updateUser(user, id).subscribe(data => {
